Add plot_map_as_groups
This function expects a map that assigns each domain element a single
group id, such that each group forms a convex set of points. This
function plots now each group as such a convex shape.

This is e.g. useful to illustrate a tiling that is given as a between
iteration vectors to tile ids.