Add function to get the faces of a 3D set
" Given a basic set we return a list of faces, where each face is
represented by a list of restricting vertices. The list of vertices is
sorted in clockwise (or counterclockwise) order around the center of the
face.  Vertices may have rational coordinates. A vertice is represented
as a three tuple. "