Add get_vertices_and_faces
Given an isl set, return a tuple that contains the vertices and faces of
this set. The vertices are sorted in lexicographic order. In the faces,
the vertices are referenced by their position within the vertex list.
The vertices of a face are sorted such that connecting subsequent
vertices yields a convex form.